+++ gpu_box_benchmark/benchmark_dockerfile_wrappers/blender_benchmark.py
@@ -17,9 +17,12 @@ from gpu_box_benchmark.numeric_benchmark_result import BenchmarkResult, Numerica
 
 LOGGER = logging.getLogger(__name__)
 
-_BLENDER_BENCHMARK_VERSION = "0.2.0"
+_BLENDER_BENCHMARK_VERSION = "0.3.0"
 """
 # Version History
+
+## 0.3.0 - (2026-05-07)
+* Bumped underlying benchmark cli version to 3.3.0.
 
 ## 0.2.0 - (2026-02-03)
 * Switched to base image w/CUDA 11.4.3 to support Kepler era cards.