+++ test/test_visualization.py
@@ -21,5 +21,6 @@ def test_create_comparison_visualization(tmpdir: LocalPath) -> None:
             map(numeric_benchmark_result.load_system_evaluation_from_disk, ALL_SAMPLE_OUTPUTS)
         ),
         output_path=Path(tmpdir) / "sample_output.png",
+        flatten_versions=True,
         title="yes",
     )